requirements.txt
Django==2.0.2
djangorestframework==3.7.7
Jinja2==2.10
MarkupSafe==1.0
pytz==2017.3
